int -> bigint

+++ db/migrate/20191002104531_add_not_null_constraint_to_foods_notices.rb
@@ -3,10 +3,10 @@
 class AddNotNullConstraintToFoodsNotices < ActiveRecord::Migration[5.2]
   def change
     change_column :foods, :amount, :float, null: false, default: 0.0
-    change_column :foods, :unit_id, :integer, null: false
-    change_column :foods, :created_user_id, :integer, null: false
-    change_column :foods, :updated_user_id, :integer, null: false
-    change_column :notices, :food_id, :integer, null: false
-    change_column :notices, :created_user_id, :integer, null: false
+    change_column :foods, :unit_id, :bigint, null: false
+    change_column :foods, :created_user_id, :bigint, null: false
+    change_column :foods, :updated_user_id, :bigint, null: false
+    change_column :notices, :food_id, :bigint, null: false
+    change_column :notices, :created_user_id, :bigint, null: false
   end
 end
