+++ db/migrate/20191002145112_add_updated_user_to_notices.rb
@@ -1,6 +1,6 @@
 class AddUpdatedUserToNotices < ActiveRecord::Migration[5.2]
   def change
     add_reference :notices, :updated_user, foreign_key: { to_table: :users }
-    change_column :notices, :updated_user_id, :integer, null: false
+    change_column :notices, :updated_user_id, :bigint, null: false
   end
 end
